requirements.txt
pandas==2.3.3
requests==2.32.3
python-dotenv==1.0.1
dbt-core==1.10.13
dbt-postgres==1.9.1
SQLAlchemy==2.0.22
psycopg2-binary==2.9.7
pyodbc==5.3.0